% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{create_synthetic_example}
\alias{create_synthetic_example}
\title{Example data for Create.Synthetic()}
\format{
a list with
\itemize{
\item{nrep}{ when generating the synthetic data, replicate the observed X nrep times}
\item{datan}{ simulated internal data set}
\item{betaHatExt_list}{ list of external model estimates}
}
}
\description{
Example data set for Create.Synthetic()
}
\keyword{data}
