% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{expit}
\alias{expit}
\title{Expit function}
\usage{
expit(x)
}
\arguments{
\item{x}{vector to expit}
}
\value{
numeric vector with the value of the expit function
y = expit(x) = exp(x)/(1+exp(x)).

Expit helper function.
}
\description{
Expit function
}
