% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{initial_estimates_example}
\alias{initial_estimates_example}
\title{Example data for Initial.estimates()}
\format{
a list with
\itemize{
\item{datan}{ simulated internal data set}
\item{gamma.I}{ internal gamma coefficients}
\item{beta}{ beta estimates from external model 1}
}
}
\description{
Example data set for Initial.estimates()
}
\keyword{data}
