% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResampleGamma.R
\name{Resample.gamma.continuousY}
\alias{Resample.gamma.continuousY}
\title{Resample for bootstrap variance continuous Y}
\usage{
Resample.gamma.continuousY(data, indices, q)
}
\arguments{
\item{data}{synthetic data}

\item{indices}{indices ??}

\item{q}{length of (X,B) in the full model, including the intercept}
}
\value{
numeric vector of regression coefficients
}
\description{
Resampling function to get bootstrap variance for continuous Y.
}
\references{
Reference: Gu, T., Taylor, J.M.G. and Mukherjee, B. (2021) Regression
inference for multiple populations by integrating summary-level data using stacked
imputations \url{arXiv:http://arxiv.org/abs/2106.06835}.
}
